RankingShap/requirements.txt
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.50.0
imageio==2.34.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightgbm==4.3.0
lime==0.2.0.1
llvmlite==0.42.0
matplotlib==3.8.3
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pyltr==0.2.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
shap==0.45.0
six==1.16.0
xai-sharp>=0.1.1
slicer==0.0.7
threadpoolctl==3.4.0
tifffile==2024.2.12
tqdm==4.66.2
tzdata==2024.1
